[fix]: 修复null string问题
