+++ instrumentation/otel-webserver-module/include/core/api/WSAgent.h
@@ -82,7 +82,6 @@ public:
 
     OTEL_SDK_STATUS_CODE endClientInteraction(
         OTEL_SDK_HANDLE_REQ reqHandle,
-        bool ignoreBackend,
         EndClientInteractionPayload *payload);
 
     /**
