+++ instrumentation/otel-webserver-module/src/nginx/ngx_http_opentelemetry_log.c
