+++ instrumentation/otel-webserver-module/src/nginx/ngx_http_opentelemetry_log.h
@@ -17,11 +17,8 @@
 #ifndef __NGX_HTTP_OPENTELEMETRY_LOG_H
 #define __NGX_HTTP_OPENTELEMETRY_LOG_H
 
-extern "C" {
 #include <ngx_config.h>
 #include <ngx_core.h>
-}
-
 #include <stdbool.h>
 #include <stdarg.h>
