+++ instrumentation/otel-webserver-module/src/nginx/script.h
@@ -1,52 +1,33 @@
-#pragma once
+#ifndef __NGX_HTTP_OPENTELEMETRY_SCRIPT_H
+#define __NGX_HTTP_OPENTELEMETRY_SCRIPT_H
 
-extern "C" {
 #include <ngx_config.h>
 #include <ngx_core.h>
 #include <ngx_http.h>
-}
-
-#include "nginx_utils.h"
+#include <ngx_string.h>
 
 struct NgxCompiledScript {
-  ngx_str_t pattern = ngx_null_string;
-  ngx_array_t* lengths = nullptr;
-  ngx_array_t* values = nullptr;
-
-  bool Run(ngx_http_request_t* req, ngx_str_t* result) {
-    if (!lengths) {
-      *result = pattern;
-      return true;
-    }
-
-    if (!ngx_http_script_run(req, result, lengths->elts, 0, values->elts)) {
-      *result = ngx_null_string;
-      return false;
-    }
-
-    return true;
-  }
-
-  bool IsEmpty() {
-    return pattern.len == 0;
-  }
+  ngx_str_t pattern;
+  ngx_array_t* lengths;
+  ngx_array_t* values;
 };
 
 struct ScriptAttributeDeclaration {
-  ScriptAttributeDeclaration(const char* attribute, const char* script) :
-        attribute(ToNgxString(attribute)), script(ToNgxString(script)) {}
-  ScriptAttributeDeclaration(ngx_str_t attribute, ngx_str_t script) : attribute(attribute), script(script) {}
-
-  ngx_str_t attribute;
-  ngx_str_t script;
+  char attribute[30];
+  char script[30];
 };
 
 struct CompiledScriptAttribute {
-  NgxCompiledScript key;
-  NgxCompiledScript value;
+  struct NgxCompiledScript key;
+  struct NgxCompiledScript value;
 };
 
-bool CompileScript(ngx_conf_t* conf, ngx_str_t pattern, NgxCompiledScript* script);
-bool CompileScriptAttribute(
-  ngx_conf_t* conf, ScriptAttributeDeclaration declaration,
-  CompiledScriptAttribute* compiledAttribute);
+int runScript(ngx_http_request_t* req, ngx_str_t* result, struct NgxCompiledScript* script);
+int CompileScript(ngx_conf_t* conf, ngx_str_t pattern, struct NgxCompiledScript* script);
+int CompileScriptAttribute(
+  ngx_conf_t* conf,
+  ngx_str_t* attribute,
+  ngx_str_t* script,
+  struct CompiledScriptAttribute* compiledAttribute);
+
+#endif
